Added some plumbing to handle session keys

+++ cpp/libs/secauthv5/src/secauthv5/AuthConstants.h
@@ -36,15 +36,15 @@ struct AuthConstants : openpal::PureStatic
 	const static uint16_t MIN_CHALLENGE_DATA_SIZE = 4;
 	const static uint16_t MAX_CHALLENGE_DATA_SIZE = 64;
 
-	const static uint16_t MIN_SESSION_KEY_SIZE_BYTES = 16;
-	const static uint16_t MAX_SESSION_KEY_SIZE_BYTES = 32;
+	const static uint32_t MIN_SESSION_KEY_SIZE_BYTES = 16;
+	const static uint32_t MAX_SESSION_KEY_SIZE_BYTES = 32;
 
-	static uint16_t GetBoundedSessionKeySize(uint16_t size)
+	static uint32_t GetBoundedSessionKeySize(uint32_t size)
 	{
 		return openpal::Bounded(size, MIN_SESSION_KEY_SIZE_BYTES, MAX_SESSION_KEY_SIZE_BYTES);
 	}
 
-	static bool SessionKeySizeWithinLimits(uint16_t size)
+	static bool SessionKeySizeWithinLimits(uint32_t size)
 	{
 		return openpal::WithinLimits(size, MIN_SESSION_KEY_SIZE_BYTES, MAX_SESSION_KEY_SIZE_BYTES);
 	}
